+++ _src/navball.cpp
@@ -1,102 +1,6 @@
 #include "navball.h"
-#include "globals.h"
 
-#include <QVBoxLayout>
-
-#include <Qt3DRender/qcamera.h>
-#include <Qt3DCore/qentity.h>
-#include <Qt3DRender/qcameralens.h>
-
-#include <QtGui/QScreen>
-
-#include <Qt3DInput/QInputAspect>
-
-#include <Qt3DExtras/qtorusmesh.h>
-#include <Qt3DRender/qmesh.h>
-#include <Qt3DRender/qtechnique.h>
-#include <Qt3DRender/qmaterial.h>
-#include <Qt3DRender/qeffect.h>
-#include <Qt3DRender/qtexture.h>
-#include <Qt3DRender/qrenderpass.h>
-#include <Qt3DRender/qsceneloader.h>
-#include <Qt3DRender/qpointlight.h>
-
-#include <Qt3DCore/qtransform.h>
-#include <Qt3DCore/qaspectengine.h>
-
-#include <Qt3DRender/qrenderaspect.h>
-#include <Qt3DExtras/qforwardrenderer.h>
-
-#include <Qt3DExtras/qt3dwindow.h>
-#include <Qt3DExtras/qfirstpersoncameracontroller.h>
-
-
-Navball::Navball(QWidget *parent) : QWidget(parent)
+Navball::Navball()
 {
 
-    Qt3DExtras::Qt3DWindow* view = new Qt3DExtras::Qt3DWindow();
-    view->defaultFrameGraph()->setClearColor(QColor(QRgb(0x000000)));
-    windowContainer = QWidget::createWindowContainer(view, this);
-
-    // Root entity
-    rootEntity = new Qt3DCore::QEntity();
-
-    // Camera
-    cameraEntity = view->camera();
-    cameraEntity->lens()->setPerspectiveProjection(45.0f, 16.0f/9.0f, 0.1f, 1000.0f);
-    cameraEntity->setPosition(QVector3D(0, 0, 20.0f));
-    cameraEntity->setUpVector(QVector3D(0, 1, 0));
-    cameraEntity->setViewCenter(QVector3D(0, 0, 0));
-    QSize navballSize(globals::NAVBALL_WIDTH, globals::NAVBALL_HEIGHT);
-    windowContainer->setMinimumSize(navballSize);
-    windowContainer->setMaximumSize(navballSize);
-
-    // Light
-    Qt3DCore::QEntity *lightEntity = new Qt3DCore::QEntity(rootEntity);
-    Qt3DRender::QPointLight *light = new Qt3DRender::QPointLight(lightEntity);
-    light->setColor("white");
-    light->setIntensity(1);
-    lightEntity->addComponent(light);
-    Qt3DCore::QTransform *lightTransform = new Qt3DCore::QTransform(lightEntity);
-    lightTransform->setTranslation(cameraEntity->position());
-    lightEntity->addComponent(lightTransform);
-
-    // For camera controls
-    Qt3DExtras::QFirstPersonCameraController *camController = new Qt3DExtras::QFirstPersonCameraController(rootEntity);
-    camController->setCamera(cameraEntity);
-
-    // Creat Navball entity
-    createNavballEntity();
-
-    // Set root object of the scene
-    view->setRootEntity(rootEntity);
-
-    QVBoxLayout* vBoxLayout = new QVBoxLayout;
-    vBoxLayout->addWidget(windowContainer);
-    setLayout(vBoxLayout);
-}
-
-
-void Navball::createNavballEntity() {
-      // Sphere shape data
-      Qt3DExtras::QSphereMesh *sphereMesh = new Qt3DExtras::QSphereMesh();
-      sphereMesh->setRings(20);
-      sphereMesh->setSlices(20);
-      sphereMesh->setRadius(2);
-
-      // Sphere mesh transform
-      Qt3DCore::QTransform *sphereTransform = new Qt3DCore::QTransform();
-
-      sphereTransform->setScale(1.3f);
-      sphereTransform->setTranslation(QVector3D(-5.0f, -4.0f, 0.0f));
-
-      Qt3DExtras::QPhongMaterial *sphereMaterial = new Qt3DExtras::QPhongMaterial();
-      sphereMaterial->setDiffuse(QColor(QRgb(0xa69929)));
-
-      // Sphere
-      sphereEntity = new Qt3DCore::QEntity(rootEntity);
-      sphereEntity->addComponent(sphereMesh);
-      sphereEntity->addComponent(sphereMaterial);
-      sphereEntity->addComponent(sphereTransform);
-      sphereEntity->setEnabled(true);
 }
